create createProject file
